Move agent runner to Articles namespace
Drop Dry::Monads from the runner and let the service wrap returned
articles or finish reasons in Success/Failure.
